+++ Public/Search-WiFiNetwork.ps1
@@ -1,6 +1,15 @@
 <#
     .SYNOPSIS
         Requests a scan for available networks on the indicated interface.
+        The scan is requested by calling the WlanScan function in the WlanApi
+
+    .PARAMETER WiFiAdapterName
+        Specifies the name of the WiFi adapter that will be used to search for WiFi networks.
+
+    .EXAMPLE
+        PS C:\>Search-WiFiNetwork WiFiAdapterName WiFi
+
+        This examples will search for WiFi netowrks on the WiFi adapter.
 #>
 function Search-WiFiNetwork
 {